Add "-d/--depsfile", disconnect from -- vs. -c
Enable *separate* control of (1) how to read commands and
(2) the output to be generated.

The program still supports either -c "command" or -- command args
as input, but now the user can select
the -o JSON format, the -d depfile format, or both as output
regardless of how the command was provided.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>