pmaudit: Use -j/--json flag, add --custom
This makes further improvements.  We can continue to
generate JSON or makefile dependencies (or both) regardless
of the command input, but we switch to -j/--json to request it.

This adds --custom KEY=VALUE to enable storing custom values
in the JSON file.

I think this responds to all the immediate comments
(other than passing on arbitrary flags, which is a separate discussion).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>