Add FINAL_CMD list information
The "CMD" entry provides a simple string display of the
executed command, this if fine for basic monitoring.
However, it doesn't show the exact parameter
boundaries, and it omits the shell and options used to invoke it;
those matter in some cases.

This commit adds FINAL_CMD that shows the full and exact
command being executed, when you need the exact details.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>