Substantially reworked and enhanced. Details:
- It still had some Python2-compatibility hacks.  Those are
gone and it's now completely reliant on 3.x. As of this release
3.7 is the lowest supported Python version.

- There used to be a bug in os.walk() affecting timestamps of
symlinks so I had to use an adapted version of the deprecated old
os.path.walk(). But I no longer see the bug in os.walk() so the old
hack is gone and os.walk() is used.

- In my own day job we have a situation where we're only interested in
analyzing what happens after a given point in the build.  The earlier
part is uninteresting setup but since it's all driven from the same DAG
that's hard to separate. The solution added here is to provide a -r,
--ref-file whose timestamp indicates the virtual build start.

- An --exclude option allowing certain file types to be ignored is added.

- All time/timestamp data is now kept in integral
nanoseconds-since-the-epoch as returned by stat().st_mtime_ns and so on.
No more floating-point seconds. The Python3 API guarantees to provide
integral nanoseconds and I assume that on a platform/filesystem with less
granularity nanoseconds are still returned, just padded with zeroes to
the right. Linux has had nanoseconds for a long time.

- Docstrings, usage message, comments are all enhanced.

- Options are cleaned up and reorganized. A couple are even renamed.
The concept of "extended help" is implemented. Extended help is requested
with --HELP instead of --help. Some advanced or rarely-used options
have been demoted to extended status.